EngineerAi/requirements.txt
flask
google-genai
python-dotenv
gunicorn
requests